Update operator helm chart to check for existence of OLM
Summary:
We've been running into issues lately where users are deploying to OpenShift (or other clusters with OLM already deployed) without the `deployOLM` flag set to false.
We can (mostly) autodetect the existence of OLM in the cluster, and can use that to determine whether we should deploy OLM or not. This is a helm-specific function, so should only be updated in the helm templates itself (hence the `sed` command).

Other caveats: the user may have OLM installed, but no operatorgroups.

Test Plan: deploy CLI with/without deployOLM. deploy with helm chart with deployOLM=true, deployOLM=false, no deployOLM flag applied with/without OLM deployed.

Reviewers: vihang, zasgar

Reviewed By: zasgar

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11151

GitOrigin-RevId: c3f1da90d432675bd6eb5685b60f3696e9671a12