[PC-1487]: Remember embedded theme in editor
Summary:
The vis editor wasn't honoring the theme, and apparently when it used its default value the PxL editor stopped checking that property too.

Making sure that both editors use the theme's palette is enough to fix both problems.

Test Plan: Embed Pixie in an iframe with the `/embed` routes. Press Ctrl/Cmd+E to open the editor. Switch back and forth between the PxL and Vis tabs in the editor that pops up in the side. It should retain the theme that's passed by the embedding file.

Reviewers: michelle, vihang

Reviewed By: michelle, vihang

Subscribers: nserrino

JIRA Issues: PC-1487

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11160

GitOrigin-RevId: cf4c801ec0c684880dabe2419e0ca45427b79930