Clean up the docstring generator to remove the shell script
Summary:
docstring generator was super messy and not necessary. This changes it to collapse the
shell script into the golang file. I kept the shell test around because it's an easy way to verify
the docstrings are formatted correctly.

Test Plan: Tests pass, made sure that hte output documentation is identical from before the change.

Reviewers: vihang, htroisi, michelle

Reviewed By: vihang, htroisi

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11196

GitOrigin-RevId: d6501e8c8cb798b4d85d81190a1e228b9e9a3935