PP-3356: Allow users to specify memory request in addition to memory limit for PEMs
Summary: Users want to be able to specify both the memory request and the memory limit for PEMs. Perhaps they only have 2Gi of memory available at deploy time per PEM, but they want to allow for a limit of up to 3Gi before OOMkilling the PEM. This diff updates the CLI, cloud, and operator in order to support that. Additionally, the table store was previously sized at 60% of the total memory limit. Now, the table store will be sized at 60% of the total memory request, to prevent OOMKills that could arise when the two values are different. By default the request will be the same as the limit. These aren't breaking changes, but the request will be the same as the limit in all cases until the user is running the latest release of the CLI, operator, and cloud.

Test Plan: ran skaffold versions

Reviewers: michelle, oazizi, vihang

Reviewed By: michelle

JIRA Issues: PP-3356

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11198

GitOrigin-RevId: 763248a88d824fa6fd53db8e5cb8bb342f793697