Update plugin gql to allow configuring Insecure TLS
Summary:
This updates the gql resolver to allow users to fetch whether insecure TLS is allowed for a plugin, and allows user to configure whether insecure TLS should be used or not.
This usecase will be common for users who stand up an OTel collector (or other OSS tool) in their cluster but don't want to deal with SSL certs/productionizing their workload yet.

Test Plan: Resolver tests succeed

Reviewers: nlanam, vihang, philkuz, nserrino

Reviewed By: nlanam

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11221

GitOrigin-RevId: 0b6a4c9b27de10d8cd6e7cfb253fddfd4f80ce1a