[PP-3371] Support manual specification of unit string value for OTel metrics
Summary:
We currently support unit values from semantic types. However the UX around this
specification is very hard to use and even if we fix it there will still be a missing feature
that OTel users will desire.

We decided to just build that missing feature: adding a way to manually specify the unit string.
Now this value will override any inferred semantic type value. Remove if you want to use the sem
type.

Test Plan: Added tests for the propagation of the unit in the otel PxL module tests.

Reviewers: michelle, jamesbartlett

Reviewed By: jamesbartlett

JIRA Issues: PP-3371

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11253

GitOrigin-RevId: f4d47d9a0505bc79eeaf7642876ed8670750356f