Fix `helm template` for OLM chart
Summary:
A previous change added a line to the OLM template to only deploy OLM if not present in the cluster.
Unfortunately, the `lookup` command is not supported in `helm template`. This diff updates the template line to do more careful checking to handle when the `lookup` is nil. Short circuiting isn't supported in templates, so this is unfortunately more verbose.

Closes #421

Test Plan: Deployed operator RC and ran `helm template` and `helm install` with/without OLM on the cluster.

Reviewers: vihang

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11254

GitOrigin-RevId: 131f6b4d5840e3c9aea0e8b90bc87688e83b0080