Add ParseProcPIDPss to system/proc_parser.h|cc to get "PSS" (proportional set size) memory usage for a given pid.
Summary: In many cases, PSS is a more useful measure of memory use (vs. RSS or VSIZE). In this diff, we add the capability to parse out the "PSS" or proportional set size for a given pid/process. The PSS is the memory usage of a process after accounting for shared pages. For example, if 10 processes map in a shared library that is 10MB in size, each process is only charged 1MB to their PSS. For RSS and VSIZE, each process is charged 10MB even though the memory is shared 10 ways.

Test Plan: Added a test case to proc_parser_test.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

JIRA Issues: PP-3307, PP-3320

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11263

GitOrigin-RevId: e2937a0f097f56849cdf32eb2835f59c5a560227