Fix OTel export error to make it more readable.
Summary:
The OTel export error would just use the integer enum value instead of showing the
string name of the enum. That was not readable and made debugging difficult.

I decided to fix that and change some wording around the error to make it easier to understand and
take action.

Test Plan: Added tests and happened to run it in skaffold deploy as well

Reviewers: michelle, jamesbartlett, vihang, nserrino

Reviewed By: jamesbartlett

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11292

GitOrigin-RevId: 19884838469e53ba2ce28313d055b7980e113f69