PP-3379: Propagate Execution Errors from PEM to Qb
Summary:
This diff propagates Execution Errors that a Carnot instance receives from any
parents up to its children. This enables errors from PEMs to be sent to the Kelvin then forwarded
to the query broker.

Test Plan:
Tested on skaffold where I deliberately broke the memory_source_node with an error.
Was able to see that error whenever I ran a query.

Reviewers: jamesbartlett, michelle, nserrino

Reviewed By: nserrino

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11293

GitOrigin-RevId: ff49916459c90af390bd9061d5be44a464c10bed