[PP-3390] Add metric for agent registrations.
Summary: Adds a metric that is a proxy for number of times agents have restarted. In reality, it measures the number of times each agent has registered with the metadata service. This number includes 1 for the initial registration, plus it increments each time an agent crashes or receives a HeartbeatNACK. So its a fairly good proxy for number of restarts.

Test Plan: Tested with diff that scrapes metrics from metadata service. Saw that metrics get to big query. Also tested by sshing into the node and killing a pem process to simulate a crash, saw that the agent registration metric in big query was incremented after that.

Reviewers: nserrino, michelle, vihang

Reviewed By: vihang

JIRA Issues: PP-3390

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11307

GitOrigin-RevId: dd6558c13da3a19286663ef5d0f141507e278823