Stirling: move file upid.h to a new location that is independent of 'bcc' or 'bpf' in its file path.
Summary: Many subcomponents of stirling depend on file upid.h (for struct upid_t). In this diff, we move file upid.h to a new location so that these dependencies do not include 'bcc' or 'bpf' in their file paths (i.e. to avoid the wrong implication that something depends on BCC of BPF).

Test Plan: Existing

Reviewers: #stirling, yzhao, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi, yzhao

JIRA Issues: PP-3382

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11313

GitOrigin-RevId: 4e52504aea2174f4b3bdfcfb40aafab041dfba2a