Abstract GRPC related config in Carnot::Create and add GRPCRouter as an argument
Summary:
Gone through a few variations of this change now. Carnot hid the details of the GRPCRouter, but
in reality it's an important part of the interface that was hidden away. That made testing
difficult because you couldn't send messages to the Carnot instance via the GRPCRouter.

Also the Interface became unweildy with the amount of config required set inside. This diff abstracts the grpc config into two structs: the Client and Server side config. Clients can expand to include some other clients as well. The Server is simply everything around the GRPCServer implementation that manages all inbound data paths for this Carnot instance.

The follow up diff uses the interface for a test.

Test Plan:
Follow up diff uses the interface for a test. Made sure this version still runs
queries in skaffold

Reviewers: nserrino, vihang

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11300

GitOrigin-RevId: 58f3af6518d05663a1e40ffbf6821c9d0ba91cac