Enhance openssl tracing to fallback to function pointers when dlsym fails
Summary:
This is a step towards adding TLS tracing support to netty.
The netty-tcnative shared library does not use a dynamic OpenSSL_num_version symbol.
This means the symbol ends up in the symtab rather than dynsym. The former is private and incompatible with dlsym, while the latter is public and is supported by the existing openssl tracing.

Test Plan: Updated openssl_trace_bpf_test to cover the new cases.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Signed-off-by: Dom Del Nano <ddelnano@twitter.com>

Differential Revision: https://phab.corp.pixielabs.ai/D11327

GitOrigin-RevId: 20abc6c7636f923e3d587e5467748397ee4c97da