Fix blocking agg benchmark
Summary:
The blocking agg benchmark was writing to the same result server for exec stats.
The exec_stats() call of the localresultsserver threw an error, causing the bm to fail.

I added a new method to reset the data stored in the localresultserver so now the bm can avoid
the error.

Test Plan: Ran the bm and it passed.

Reviewers: vihang, nserrino, jamesbartlett

Reviewed By: jamesbartlett

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11324

GitOrigin-RevId: 7a0999088409e931464b97d250f542bb3d256aea