PerfProfiler: improve the test app Fib.java such that newer JVMs do not optimize it away.
Summary: Fib.java is a toy app designed to produce a predictable workload such that our performance profiler can reasonably expect to find one symbol twice as often as another. Succinctly, we believe the profiler is working if we run Fib.java and meet this expectation. But, for newer JVMs, our previous version of Fib.java was optimized such that we found each symbol at a ratio of 1:1 (vs. 2:1, the expectation). In this diff, we make our toy app harder to optimize and thus enable its use as a test application with newer JVMs.

Test Plan: Existing.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

JIRA Issues: PP-3373

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11329

GitOrigin-RevId: a77194a76b69a5c5325159c9e5503c641e5b82fb