Add test for UpstreamConnectionTimeouts and GRPCBridge for Carnot
Summary:
After some of the error changes, Natalie brought up a case where we didn't know what
happened whether a Parent dying in an Agent graph would cause an error to be thrown down stream.
We currently show this data to the user even if it's not complete.

This adds a test to make sure that the Execute call successfully completes and downstream consumers
don't receive an error.

Test Plan: Added tests; no functional changes.

Reviewers: nserrino, jamesbartlett

Reviewed By: nserrino

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11331

GitOrigin-RevId: cfc01b5d7ad49c83da94c3064bc04d06d04fee2f