PP-3394, pt1: Add InitiateConnection to TransferResultChunk message
Summary:
We need to support sending Errors before a GRPCSinkNode can
call Open(). There are bound to be cases where this happens. When I added a test for this case
in carnot_test, the GRPCRouter complains that the query has already been deleted. That's because
GRPCRouter expects a TransferResultChunk request to initiate a stream for one of the GRPCSource
nodes. However, if Open() hasn't been called yet on the error-ing agent side, this message will
never be sent. It makes more sense to move this open logic up to the query level and then
have `carnot.cc` make this call instead.

The design I've settled on is adding an InitConnection that will be made at the
start of each query to all the outgoing_connections. GRPCRouter will now look for this instead
of an init call. The init call will be deprecated and replaced.

This diff is the interface change.

Here's the Implementation plan:
1. InitiateConnection handling in QueryBroker that roughly duplicates initiate_stream_results
2. In Carnot: Send over InitiateConnection,  ignore incoming initiate_stream_results,
    stop sending outgoing initiate_stream_results
3. QB: Ignore initiate_stream_results
4. Deprecate initiate_stream_results and remove all support in QB and Carnot

Alternatives Considered and why they aren't in play:
1. Make the GRPCSink nodes call Open() first
  * hacky, we don't need to expose this impl detail to Carnot.
2. Call the initiate stream calls for all grpc_source_nodes
  * similar to the above. It kind of abuses the contract that the interface of initiate_stream_results implies. I don't think we need the interface anyways so we can drop it.
3. Allow errors to arrive without having a query
  * Hard to tell if query is already deleted or is new. Potential to lead to a memory leak, the same kind that James fixed about a year ago.

Test Plan:
The entire change will be tested using the exectime_benchmark pointing to a dev version of Vizier running this change.

That will tell us if any scripts were disabled after the change and test the change against a few hundred script executions in case something flakey occurs.

This specific diff is untested because it's an interface addition without any users.

Reviewers: nserrino, jamesbartlett, vihang

Reviewed By: jamesbartlett

JIRA Issues: PP-3394

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11333

GitOrigin-RevId: 628c07fbb45e2d950f32123118219b6e66022693