GRPC-C Library Tracing Support: Prep-work
Summary: This is a step towards adding support for tracing GRPC traffic through the GRPC-C library. This diff includes the event handlers on the user space side.

Test Plan:
This functionality is not enabled. Test should be added before this is turned on for.

Existing tests ensure the current changes don't break anything.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Signed-off-by: Ori Shussman <ori@groundcover.com>

Differential Revision: https://phab.corp.pixielabs.ai/D11875

GitOrigin-RevId: f18fc19291c4ef7b5b71138d09d525e6800f1b3d