PP-3394, pt5: Deprecate initiate_result_stream
Summary:
This diff deprecates the interface and removes it from the system. The largest functional
change is the removal of checking for an initiate_result_stream call. That now leads the sinks
to send out a zero-row row batch over grpc. That caused the most line changes in the tests.

Test Plan: exectime on skaffold'd dev env

Reviewers: jamesbartlett, michelle, nserrino

Reviewed By: jamesbartlett

JIRA Issues: PP-3394

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11342

GitOrigin-RevId: 0b68d6f8f3357e84d9d09b1434b9c0d9459d318a