Fix bug where nodeselector overwritten if specified in CRD `patches`
Summary: Users can specify a nodeselector using the `patches` field. It should behave like a normal `kubectl patch`. However, we also added supported for `nodeSelector`s down the line, and this ended up overwriting the nodeSelector specified by a patch.

Test Plan: Deploy a skaffold version of operator and a CRD with `patches`

Reviewers: vihang, nserrino

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11337

GitOrigin-RevId: 51460acb7d59a64acf3f8d5f5e86ea911a1d2b18