Operator should reconcile when CRD spec is updated
Summary:
Our operator was previously only reconciling on Vizier version changes. The reconcile phase is triggered anytime the CRD is updated (whether it is a status or spec change). However, we only want to trigger a reconcilation on any spec changes. To maintain feature parity, we made the operator only redeploy on Vizier version changes. However, this can lead to confusion if a user is manually updating the CRD to adds labels... And nothing happens.
There are a few ways to detect whether reconcilation must be performed:
- Check the spec, and the actual state of the cluster (for example, checking all pods in the cluster, the labels on the pods in the cluster, etc). This is most correct and comprehensive... But very time consuming and difficult to capture all cases we specify in our `spec`.
- Replicate the spec in the status. The status represents the state of the cluster.
- Represent the spec as a "checksum" in the status. If the spec checksum changes, that means reconcilation should be performed.
Opted for the last option here. By relying on `kubectl apply` functionality under the hood, this works as expected. For example, changing just `pemMemoryLimit` will only redeploy the PEMs since the YAMLs for the other pods remain the same.

Test Plan: Skaffold deployed operator and vizier CRD. Confirm updates to CRD leads to correct redeploys of Vizier with the newly specified options.

Reviewers: vihang, nserrino, philkuz

Reviewed By: vihang, philkuz

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11343

GitOrigin-RevId: b46c8cd0e71da1cd517d5382287431db2d4f5af5