PP-3394, pt.4: Remove the initiate_result_stream call in grpc_sink_node
Summary:
This removes the caller of initiate_result_stream from Carnot. Now this job is taken
over by the InitConnection() call inside of Carnot::Execute(). The behavior is roughly
equivalent to before. Instead of sending over the Initiate_result_stream call we send over
an empty rowbatch of 0s. This would already happen if the GRPC connection died in the middle
anyways; so this removes a bit of complexity from the path.

Test Plan: Tested on skaffold and made sure tests pass with the new behavior.

Reviewers: jamesbartlett, nserrino, michelle

Reviewed By: jamesbartlett

JIRA Issues: PP-3394

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11344

GitOrigin-RevId: 223cb18fa7348e597a87540cec8802999a567ecc