[CLEANUP] Remove all leading records with COMMAND
Summary:
For some reasons I have not investigated yet, on some host, like enigma
and Jenkins workers, redis-cli would sent 2 commands at the beginning to the server:
COMMAND DOCS
COMMAND

This does not happen on my laptop.

COMMAND DOCS accepts one argument, which is the name of a command, and
the command returns a specialized documents on the supplied command.

Since we are interested in those commands, we'll just erase all of them.

Test Plan: Existing test

Reviewers: #stirling, jps

Reviewed By: #stirling, jps

Signed-off-by: yzhao1012 <yzhao@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11345

GitOrigin-RevId: b1a0a721c43d553ca2257f09dafe2ceb13df51a0