PerfProfiler: adding java base images for test coverage.
Summary: We add 8 new base images with various JDKs to expand test coverage for Java symbolization. So that tests run in a reasonable amount of time, we reduce the profiler test instance run time from 90 seconds (per test) to 15 seconds. To support this change, we also tweak various profiler defaults as necessary: we reduce the table push period from 30 sec. to 5 sec. and we increase the sampling freq. to reduce variance in our results.

Test Plan: This is a test and it expands our test coverage.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11175

GitOrigin-RevId: 5e53a9c47902c60566edea6a448d26262784d983