Add timestamp partitioning to vizier metrics bigquery table.
Summary: This adds timestamp partitioning to the table so that we don't pay for querying all of the data everytime. Also changes the name of the table so that we can still access the current data for now, and delete the table later.

Test Plan: Tested on testing cluster, creates the new table correctly.

Reviewers: vihang, michelle, nserrino

Reviewed By: vihang

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11354

GitOrigin-RevId: 2e4720baaec05eb1c8ffa708a433c0ca45e82839