[PC-1508]: Better tooltips and toasts for plugin enablement
Summary:
Enabling and disabling plugins has consequences that can be surprising.

There was already a tooltip for disabling, but not for enabling. Neither had a toast to confirm what the user had just done. Now both actions get a tooltip and a toast afterwards. The toast after enabling a plugin also has a call to action to configure the newly-enabled scripts that come with it.

Test Plan: Enable and disable a few plugins. There should be a different tooltip depending on whether you're enabling or disabling, on the checkbox. Additionally, both actions trigger a toast.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-1508

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11356

GitOrigin-RevId: 5b9959eb4f2ebb9e865b25cd2835bc96a6c32ba6