Update CLI api-key/deploy-key create to include -s
Summary: Some users may want to pipe the output of api key/deploy key creation to other tools. In the current state, there is a bunch of additional text (such as id). This diff adds a `-s` for `short` so that the create command only returns the key.

Test Plan: Build and run CLI `px api-key create -s` and `px deploy-key create -s`

Reviewers: vihang, philkuz

Reviewed By: vihang, philkuz

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11359

GitOrigin-RevId: 2382cf0f5ca487545ace9cc867e5ca379508df1a