Add `px get cluster` to get more information about cluster in current kubeconfig
Summary:
Users may have Pixie running on a cluster, but it may be difficult to identify which cloud the cluster is connected to and what is the cluster ID.
Added a `px get cluster` command to get that info. by passing `--id` or `--cloud-addr`, the user can isolate which fields they may want to easily pass to other tools.

Test Plan: Build CLI and run `px get cluster`

Reviewers: vihang, philkuz, nserrino

Reviewed By: philkuz

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11361

GitOrigin-RevId: a8845a29f83a11c7d8a69b8a0512782f98789e65