Add CIDR UDFs.
Summary:
Adds two UDFs, one for getting the set of CIDR ranges that represent pods and services in the cluster.
The second checks if a given IP is contained in any of the given set of CIDR ranges.

These UDFs are useful for the data exfiltration demo.

Test Plan: Added test for get_cidrs. Tested both as part of the demo.

Reviewers: nserrino, vihang, philkuz, zasgar

Reviewed By: nserrino, vihang

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11383

GitOrigin-RevId: a01b52b9cd0fb8884d0b83a3fdc1727d7110aeb9