Track number of PEMs queried per script.
Summary: Adds a prometheus summary for the number of PEMs queried for each script. A future diff will add a metric for the number of PEMs that returned results for each script.

Test Plan: Tested by skaffolding to a cluster connected to testing, and made sure the metric shows up in big query.

Reviewers: nserrino, philkuz, vihang

Reviewed By: vihang

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11390

GitOrigin-RevId: 8d8061a30dfa50fa56edf6a7641f7bfce473c974