Batch writes from metrics service to BQ
Summary:
The mertics NATS consumer is lagging, my guess it that it is due to
the fact that we make a write per timeseries instead of batching them. This
should help a bunch. Eventually we could decouple the nats reading from the bq
write and run them in separate routies.

Test Plan: Deployed to testing cluster.

Reviewers: jamesbartlett, michelle

Reviewed By: jamesbartlett, michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11418

GitOrigin-RevId: 6b12867619e9ec57af5d554f6c341647de6cab69