[PP-3398] Add compiler type checking for Join index columns.
Summary: Previously, kelvin would crash if you misspecified a Join in PxL. This diff adds a type check in the planner to make sure that the join on columns have the same types in both parents.

Test Plan: Added a unit test. Also, tested in skaffold with the previously crash inducing script, which no longer crashes but instead returns a compiler error.

Reviewers: philkuz, nserrino

Reviewed By: philkuz

JIRA Issues: PP-3398

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11422

GitOrigin-RevId: cb15e40d15858b16269904426bf6df79c498699b