Add timeouts to metrics BQ put
Summary:
BQ put retires and the accompanying context should include a timeout
to prevent retries from hanging forever. This diff also switches over to using
StructSavers with a schema to prevent a bunch of calls to reflect and Schema
inference within the BQ library.

Test Plan: Deployed to testing cluster.

Reviewers: jamesbartlett, michelle

Reviewed By: jamesbartlett

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11423

GitOrigin-RevId: 63409394ae13e63368e661f6fd786c779b68929e