[PP-3114] Fix column renaming bug in filter push down rule.
Summary: The filter push down rule had a bug where if the filter had already been pushed past another op, and there was a map it was attempting to push past, with 2 columns, one a simple rename and the other a calculated column, then the rule would first update the filter's column names based on the map's rename, and then realize that it couldn't push the filter down past the map. However, it doesn't reverse the column name change after realizing it can't push down past the map, leading to "column does not exist in TableType" errors.

Test Plan: Added tests for the bug, that failed before and pass now. Also tested that the query in the JIRA works on a skaffolded vizier.

Reviewers: nserrino, philkuz, vihang

Reviewed By: philkuz

JIRA Issues: PP-3114

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11427

GitOrigin-RevId: 47010cd6c5fc9229c55f808087c087807c2d9a5c