Fix default passthrough port on local builds
Summary: On local builds, the default value for the passthrough port was an empty string. This propagated through to `flags.js` as a direct replacement, which became a syntax error. Making it propagate as an empty string in JS syntax fixes it; this change also makes sure to pull the value from the domain YAML when possible, instead of ignoring it.

Test Plan: `yarn dev`. Try to load the page.

Reviewers: michelle, vihang

Reviewed By: vihang

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11440

GitOrigin-RevId: 5f5a14dfe4358f54f8d72794db6c6993efbc8e4f