[PC-1531]: Improve theme construction logic
Summary:
Makes sure that theme augmentations (the ones not directly defined by the palette) follow the given palette rather than the default one.

For example, if an embed sets a custom primary color, that color will now propagate to default link styles (it previously didn't know to do that).

Test Plan: Load the local UI normally and in an embed. Try the light theme. Everything should still look the same, with the exception of "default" styled links (`<a>`). Those should now reflect the primary color even when embedding with a custom theme.

Reviewers: michelle, vihang

Reviewed By: vihang

JIRA Issues: PC-1531

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11443

GitOrigin-RevId: e7937fdd1211caccc284c513279a389c510df6a3