PP-3408 - Update timer to remember to call t.Stop() in query broker
Summary: We are seeing excessive CPU spikes / garbage collection in the query broker every 2 minutes, which corresponds to the natsWaitTimeout. It turns out that we aren't calling t.Stop() on the timer, which is an important step. Let's add that in to see if it addresses the CPU spikes.

Test Plan: run it and see if it helps CPU/GC

Reviewers: michelle, zasgar, vihang

Reviewed By: vihang

JIRA Issues: PP-3408

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11457

GitOrigin-RevId: f26d431be7c0a3a0fc56b0337cba07db888c8f7a