[PC-1535]: Enable embedding the Data Retention view
Summary:
Makes it possible to embed the URL `/embed/configure-data-export` and configure scripts for enabled plugins.

Changes the visible messaging a little bit while in this view and hides the nav bars, much like embedding `/embed/live` does, so that you can't escape the widget into a bad state.

A bit of refactoring was involved to make this possible, but nothing too awful.

Test Plan:
Embed the above URL and make sure it works both with plugins active and with no plugins active. Should still be able to edit and create scripts as normal, but the link to `/admin/plugins` and the navbars shouldn't appear.

In the normal page (not embedded), nothing should change at all.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-1535

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11462

GitOrigin-RevId: b11db7f320621f4a6ca5c2bf0e5c53d4d1ed25eb