Improve theme selection logic again
Summary:
Embeds are tricky.

The various ways to set the theme try to fight each other without this change. With it, the priority is thus:
- `postMessage` (for either a builtin or custom theme)
- `?theme=light` URL param
- `?customTheme=url-encoded-json` URL param
- User preference only if not embedded
- Default (dark theme) if none of the above change the theme

The logic to prevent a collision is a bit gnarly,

Test Plan:
Try embedding all four ways above. This is nontrivial, instructions will be coming for future tests in a later diff so we can check all of the ways embeds happen easily.

Try running the UI locally outside of an embed as well.

In the case where you have Pixie set to `SYSTEM` for its theme, the embedded versions with an overridden theme should NOT honor it and should continue to use whatever URL parameter or postMessage they were told to by the embedding page.

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11495

GitOrigin-RevId: 48a1086b3aff6e127f7611106d8c211dd0b43ba1