Stop the `cancelled` errors
Summary:
We use cancellable promises in a few places for asynchronous tasks.

The implementation for this had a way to silence the rejection message when cancelled in this way, but it wasn't working properly.

It turned out every call site already block-scoped these promises and ignored the cancellation error, so making the default behavior simply silence the promise was the cleaner approach.

There's still an optional parameter to `makeCancellable` if a future consumer DOES need the promise to instantly reject when cancelled.

Test Plan: Try embedding, logging out and in, cancelling a long-running script, using the autocomplete, etc. You should no longer see `cancelled` as an error in the browser console. Cancelled promises should still get cut off so that any `then()` effects on them never fire.

Reviewers: michelle, vihang

Reviewed By: vihang

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11543

GitOrigin-RevId: d7d2f601739a3631921d2ea3692df35178958fb9