Add end_time_ns argument for plugin config
Summary:
adds the end_time_ns argument for plugin config and pipes it up to the final user,
the compiler. Also makes px.plugin.end_time a valid call that will actually return the end_time.

This does not change anything immediately as we have to update the scripts still.

Test Plan: Tested on skaffold and could successfully get the expected window of data

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11555

GitOrigin-RevId: dd1650bdfbee33069f1c6adcdc06eac4224747b2