Pass along sucess state of plugin config when saving
Summary:
Error check was relying on something that wasn't passed through.

I think this was a mistake when rebasing at some point that got missed.

Test Plan:
Try editing a plugin's config and saving it.
Before, it would save but tell the user that something failed. Now, it correctly tells them that the save was successful.

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11557

GitOrigin-RevId: dbbb42e6cfdd44a94badb0b21c52f295e772e083