Adjust plugin scripts to run 1s in the past
Summary:
In the existing plugin script runner we would be at the mercy of the script execution
latency as well as data collector latency. We would run scripts that would get the execution
period of data in the past up until the current moment.

Those latencies meant the plugin could oversample data by sampling across windows and undersample
data if the data for the window was not yet propagated from the collector to the table store.
This fixes the problem by sampling windows 1s in the past which solves the undersampling
issue.

Test Plan: Tested on skaffold with a custom script. Was able to see data in my otel destination.

Reviewers: michelle, vihang

Reviewed By: michelle, vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11558

GitOrigin-RevId: a26c7e07220b0147e0ac66579507abcc79490574