In dns_flow_graph, localhost IPs are resolved to pod names
Summary:
The purpose of this change is to avoid directing all call graph edge to 127.0.0.x IP address, which always leads to the loopback address of a pod, and cannot be resolved into a meaningful hostname or pod name.

With this change, a pod's message sent to a loopback address is attributed to its own pod, and to distinguish from the pod itself, we use `localhost:<pod_name>`.

Test Plan: Manual testing on webUI

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: rcheng, oazizi

JIRA Issues: PP-3416

Signed-off-by: yzhao1012 <yzhao@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11563

GitOrigin-RevId: dc1bd9a67ec67885c5df9744560ff45255dc517c