Add segment events for plugin in cloud backend
Summary:
There are already segment events for modification to the plugin/retention scripts in the frontend. However, in the future, we may want to allow plugin configuration through other clients, such as the CLI.
This diff augments the cloud backend with additional segment events to track plugin use.

Test Plan: Deploy to a testing cloud and check segment debugger to ensure the events show up.

Reviewers: vihang, nserrino

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11568

GitOrigin-RevId: 5e2d49e801b8ab5ad89d0b614c7d2164f5b88acf