Remove stirling docker build containers.
Summary:
The stirling docker containers for building specific go versions slow down builds especially when deleting the cache often (as was the case during the clang14 upgrade work). This diff removes the need for the build containers, by allowing building go binaries with different versions of the go SDK.

To do this, the diff patches rules_go to allow specifying `gosdk` on golang build rules. Then all the stirling container builds are changed to just use go_binary/go_image instead of containerized builds.

Note that it required some changes to the hardcoded offsets in some of the tests. I was worried that maybe my changes didn't work fully or there were significant issues with the new way of building things, so I looked into the differences between the binaries. The byte-by-byte differences are summarized here: https://phab.corp.pixielabs.ai/P197. The differences seems to be entirely (other than a few unaccounted for bytes) due to the fact that rules_go strips out certain build information as well as local paths from the built binaries in order to produce the same binary every build. This causes symbol offsets to be slightly offset (anywhere from 90 to 4096 bytes depending on where in the binary the symbol is), hence the changes to the hardcoded offsets in the tests.

Test Plan: Ran tests. Relying on jenkins for bpf tests.

Reviewers: #stirling, vihang, zasgar, #third_party_approvers

Reviewed By: vihang, zasgar, #third_party_approvers

Subscribers: oazizi

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11586

GitOrigin-RevId: 97f8b973ae78e2fb854a8cbcda9ed50e4675532f