In proc exit connector, only count events as a crash if they have SIGABRT or SIGSEGV.
Summary: For the proc exit connector, we previously defined a crash as any exit with non-zero exit code. In this diff, we refine that. We extract the exit signal and count as a crash only for SIGABRT, SIGSEGV, SIGILL, SIGBUS, and SIGFPE.

Test Plan: Tested one off using a custom build and skaffold. Tested with old version and extra logging, and with this diff. Observed expected outcome.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11576

GitOrigin-RevId: f7e344266e171913e47cb8d38fcc8e47455bf2b0