[PC-1540]: Show disconnected clusters that were already selected in export scripts
Summary:
Previously, a disconnected cluster would be filtered out even if it was selected before.

Now, it still appears in both the overview table and the detail view.
In the detail view, the cluster options (and selected chips) all show their status icons.
Clusters that are disconnected still appear as options if they were already selected when the detail view opened.
Other disconnected clusters are still filtered out.

{F187754}

Test Plan:
Go to `/configure-data-export`. Look for a script that is configured to use a currently-disconnected cluster.
Edit it. The disconnected cluster should still appear. If you remove it from the list and try to add it back before saving, it should work.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-1540

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11584

GitOrigin-RevId: 8d0c778d5791ca141446c7a2a5da85a0719a4523