Fix button/link error
Summary:
We had a Link with its component set to Button, which causes (harmless) invalid properties.

Flipping it to be a button with its component set to Link prevents the error; the link still works.

Test Plan:
Load `/admin/plugins`. No more console error if you have react-dev-tools.
Should still have an `Edit Scripts` link in each accordion, which should still be an `<a>` which still goes to `/configure-data-export` properly.

Reviewers: michelle, vihang

Reviewed By: michelle

Subscribers: philkuz

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11615

GitOrigin-RevId: eb9da5fa4a273b9cac0feed02d37f62f38d6c348