[PP-3426]: Show status icon of cron history in details
Summary:
The error column already has this information, but now there's an icon on the left to make it easier to notice.

This involved adding a new feature to the data tables.

{F188519}
{F188520}

Test Plan:
Visit cron history for any given script.

Try those with successful runs and those with unsuccessful runs to see both versions.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PP-3426

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11667

GitOrigin-RevId: 42ee20e58cdeea3e41b1f3e1dd5fec49a7d35da6