Remove default values for stirling config flags.
Summary:
We had duplicated the default value for stirling config flags (datastream_buffer_size and datastream_buffer_spike_size) in a lot of places.
This diff removes all the additional places the default is specified such that the source of truth is now only the default specified in the stirling code.

Test Plan: Using my diff to cut a dev release, I tested that a dev release with the new templated yamls, deploys successfully without the datastream parameters specified on the PEMs. Also tested that passing the flags to the CLI ends up with the correct flags going to the PEM.

Reviewers: michelle, vihang

Reviewed By: vihang

Subscribers: oazizi

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11703

GitOrigin-RevId: b7962bff9117ebe7fe39ce5259a41eab04c61785