Remove podSecurityPolicy
Summary:
PodSecurityPolicies have been deprecated as of k8s 1.21, and will be removed in K8s 1.25.
We currently include the podSecurityPolicy by default. Most of the time it is ignored, as people need to enable the admission controller. When the admission controller is enabled, the current podSecurityPolicy isn't correct anyway, as we would also need podSecurityPolicies for the operator. In most cases, users have needed some customization around the podSecurityPolicy to best fit their configurations anyway.

Fixes #371

Test Plan: built Vizier YAMLs

Reviewers: philkuz, vihang

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11680

GitOrigin-RevId: 17a279c94575bf83e403ee0383db9e6314b8dcff