Disable custom export scripts when disabling plugin
Summary:
Currently when a user disables a plugin, we delete all preset and custom scripts. This can be particularly annoying for custom scripts if a user has spent some time writing it.
This diff updates the plugin disabling process so that custom scripts are not deleted, just disabled.

Note that these scripts are not listed in the UI until the plugin is re-enabled. We can show them so that users may still view their scripts, however this will require some more UI work. For example, to not let the user reenable the script if the plugin is disabled

Test Plan: Unit test, deploy to testing cloud and disable/renable a plugin.

Reviewers: philkuz, vihang

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11686

GitOrigin-RevId: 52aa3158047c3a9e670f8fb36a02ecae636909ee