Don't count Cypress as embedding
Summary:
Cypress puts the application under test in an iframe. It tries to mask this, but doesn't do a perfect job and our embed detection happened to notice.

This makes an exception to the rule by explicitly checking for Cypress.
I tried having it also check that the parent frame has a Cypress-specific location (so that we can detect if we're in a nested frame for actual embed tests), but Cypress fakes the top URL so we can't detect that difference.

Test Plan: Run normally. Test that embed still behave as they should. Cypress should now load again (it was broken).

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11699

GitOrigin-RevId: 5db68c7bd1470ce8df95c1abde5c5ce3ac946d14