Update operator to detect NATS version and upgrade it if need be
Summary:
This allows the operator to upgrade the NATS vizier dep. The operator
checks the image and compares them to upgrade NATS. Note that this doesn't do
a semver check but just checks the entire image name instead.

Test Plan:
Update nats to use an old version. skaffold the operator. update
the vizier resouce to trigger a reconciliation. Ensure that nats gets upgraded.

Reviewers: michelle, philkuz, zasgar

Reviewed By: philkuz

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11777

GitOrigin-RevId: a1db2be9f80792a97937d43ad4c198bb0a3e9192