Simplify sudo_bazel_run.sh.
Summary:
In this diff, we vastly simplify sudo_bazel_run.sh. We take advantage of `--run_under=sudo` to directly use `bazel run` and `sudo` together. This re-aligns our code with the original intent of the script, and vastly simplifies our code.

As a test case demonstration, we also comment out the "helper" functions inside of the perf-profiler test. These helper functions were necessary for the previous version of the script. Maintaining the helper functions has been a source of overhead.

Test Plan: This is test collateral.

Reviewers: #stirling, oazizi, yzhao

Reviewed By: #stirling, oazizi, yzhao

Subscribers: yzhao, oazizi, rcheng

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11658

GitOrigin-RevId: 870b804eb79cddedfe652ba306821f863b5da490