Show export status rather than cluster status in history selector
Summary:
Previously, the connection state of the cluster was shown next to its name in the selector.

However, the user wants to know which clusters the current retention script is having issues on, rather than which clusters are healthy/unhealthy in general.

Now, the icon is based on a summary of the same history that is shown when that cluster is selected.

{F189764}

Test Plan: Go to `/configure-data-export`. Open the history of a script. Look in the cluster selector. The status icon of each cluster shown will populate based on the total pass/fail rate of this retention script on that cluster.

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11783

GitOrigin-RevId: 8fd661a41bf5cd247cddd2a9d30c9e0104ad9084