Update operator warning message
Summary:
The current operator warning message makes it sound like users /must/ deploy to the operator version of Pixie. This is not the case, however, we still want to notify users since it can provide a better experience.
This diff updates the message with this information.

Test Plan: Skaffold deploy, delete CRD, and verify the message shows up.

Reviewers: vihang, nserrino, htroisi

Reviewed By: htroisi

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11794

GitOrigin-RevId: 7ca5f109b2790b510c2b985ab23757f042b6717c