Include list of images in vizier tar for easy airgapped deploys
Summary:
We are in the process of making airgapped deploys easier. This diff updates the vizier_yamls tar to include the list of required images that need to be pulled, and the secrets YAMLs that need to be deployed alongside Vizier.
Now, users only need to pull this tar file to get all the information they need to deploy to an airgapped cluster, rather than pulling from multiple sources/generating the image lists themselves.

Test Plan: Build an rc, check the generated tar contains the expected contents. Deploy vizier on the new rc version, update old vizier to new rc version, to ensure changes to the tar haven't broken our usual deployment/update processes.

Reviewers: vihang, zasgar, htroisi

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11800

GitOrigin-RevId: 062b4264130020b90cb5625cd5ddf91abb796cf3