Initiate gRPC event metadata to zeros
If we don't initialize eBPF heap data to zero before each event, we are
prone to hard-to-monitor errors, because data from old events may be a
part of newer events.
Initiating the entire heap to zeros for each event is too costly, and
may cause larger LLVM compilation times.
This commit is a trade-off-solution: Initiate only relevant parts of the
event to zeros. Only the "metadata" of the event is zeroed out, which
may contain data such as the pid, timestamp, length etc. Also, when
needed, initialize the first character to a null-terminator so the
string is "empty" if not edited later on.

Signed-off-by: Ori Shussman <ori@groundcover.com>