Propagate PERCENT semantic type through Sum UDA
Summary:
I needed to write a script that summed up a column with ST_PERCENT. We don't propagate
it for some reason, so added it here. I think it makes sense to add percents together. It should
be on the user to determine whether that's logically correct. In the case of summing up CPU pct
it is logically correct of a process it is logically correct.

Test Plan:
Tested on skaffold with a script that sums up percentages. PERCENT gets propagated
through the sum UDA.

Reviewers: vihang, michelle, nserrino, jamesbartlett

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11868

GitOrigin-RevId: b026f58fda946c0e311dd38399fb74b40adf2c4f